+++ src/main/java/link/myrecipes/api/dto/exception/DefaultExceptionInfo.java
@@ -5,18 +5,18 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
-import java.time.LocalDateTime;
+import java.util.Date;
 
 @Getter
 @NoArgsConstructor
 @ToString
 public class DefaultExceptionInfo {
-    private LocalDateTime timestamp;
+    private Date timestamp;
     private int status;
     private String message;
 
     @Builder
-    public DefaultExceptionInfo(LocalDateTime timestamp, int status, String message) {
+    public DefaultExceptionInfo(Date timestamp, int status, String message) {
         this.timestamp = timestamp;
         this.status = status;
         this.message = message;
