+++ src/main/java/link/myrecipes/api/exception/CustomExceptionHandler.java
@@ -1,7 +1,7 @@
 package link.myrecipes.api.exception;
 
 import link.myrecipes.api.dto.exception.DefaultExceptionInfo;
-import link.myrecipes.api.dto.exception.ValidationError;
+import link.myrecipes.api.dto.exception.Error;
 import link.myrecipes.api.dto.exception.ValidationExceptionInfo;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
@@ -13,10 +13,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
-import java.time.LocalDateTime;
-import java.util.Collections;
-import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Date;
 
 @RestControllerAdvice
 @Slf4j
@@ -25,34 +22,16 @@ public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
     protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex, HttpHeaders headers, HttpStatus status, WebRequest request) {
         HttpStatus httpStatus =  HttpStatus.BAD_REQUEST;
 
-        List<ValidationError> validationErrorList = ex.getBindingResult()
+        Error[] errors = ex.getBindingResult()
                 .getFieldErrors()
                 .stream()
-                .map(fieldError -> new ValidationError(fieldError.getDefaultMessage(), fieldError.getField()))
-                .collect(Collectors.toList());
+                .map(fieldError -> new Error(fieldError.getDefaultMessage(), fieldError.getField()))
+                .toArray(Error[]::new);
 
         ValidationExceptionInfo validationExceptionInfo = ValidationExceptionInfo.builder()
-                .timestamp(LocalDateTime.now())
+                .timestamp(new Date())
                 .status(httpStatus.value())
-                .validationErrorList(validationErrorList)
-                .build();
-        log.error(validationExceptionInfo.toString());
-
-        return new ResponseEntity<>(validationExceptionInfo, httpStatus);
-    }
-
-    @ExceptionHandler(CustomValidationException.class)
-    public ResponseEntity<ValidationExceptionInfo> handleCustomValidationException(CustomValidationException ex) {
-        HttpStatus httpStatus =  HttpStatus.BAD_REQUEST;
-
-        List<ValidationError> validationErrorList = Collections.singletonList(
-                new ValidationError(ex.getErrors()[0].getDefaultMessage(), ex.getErrors()[0].getField())
-        );
-
-        ValidationExceptionInfo validationExceptionInfo = ValidationExceptionInfo.builder()
-                .timestamp(LocalDateTime.now())
-                .status(httpStatus.value())
-                .validationErrorList(validationErrorList)
+                .errors(errors)
                 .build();
         log.error(validationExceptionInfo.toString());
 
@@ -64,7 +43,7 @@ public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
         HttpStatus httpStatus =  HttpStatus.INTERNAL_SERVER_ERROR;
 
         DefaultExceptionInfo defaultExceptionInfo = DefaultExceptionInfo.builder()
-                .timestamp(LocalDateTime.now())
+                .timestamp(new Date())
                 .status(httpStatus.value())
                 .message(ex.getMessage())
                 .build();
@@ -76,18 +55,5 @@ public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
     @ExceptionHandler(UsernameNotFoundException.class)
     public ResponseEntity<DefaultExceptionInfo> handleUsernameNotFoundException() {
         return new ResponseEntity<>(new DefaultExceptionInfo(), HttpStatus.OK);
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<DefaultExceptionInfo> handleException(DefaultExceptionInfo ex) {
-        HttpStatus httpStatus =  HttpStatus.INTERNAL_SERVER_ERROR;
-
-        DefaultExceptionInfo defaultExceptionInfo = DefaultExceptionInfo.builder()
-                .timestamp(LocalDateTime.now())
-                .status(httpStatus.value())
-                .message(ex.getMessage())
-                .build();
-
-        return new ResponseEntity<>(defaultExceptionInfo, httpStatus);
     }
 }
