+++ src/main/java/link/myrecipes/api/exception/CustomValidationException.java
@@ -1,16 +1,15 @@
 package link.myrecipes.api.exception;
 
+import link.myrecipes.api.dto.exception.Error;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.io.Serializable;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
 public class CustomValidationException extends RuntimeException {
     private final Error[] errors;
 
     public CustomValidationException(String defaultMessage, String field){
-        this.errors = new Error[]{new Error(defaultMessage, field)};
+        this.errors = new Error[] {new Error(defaultMessage, field)};
     }
 
     public CustomValidationException(Error[] errors) {
@@ -19,23 +18,5 @@ public class CustomValidationException extends RuntimeException {
 
     public Error[] getErrors() {
         return errors;
-    }
-
-    public static class Error implements Serializable {
-        private String defaultMessage;
-        private String field;
-
-        public Error(String defaultMessage, String field) {
-            this.defaultMessage = defaultMessage;
-            this.field = field;
-        }
-
-        public String getDefaultMessage() {
-            return defaultMessage;
-        }
-
-        public String getField() {
-            return field;
-        }
     }
 }
